My Work So Far
Linking from a tile on home pulls documents under the category.

Linking from a tile on home sends the user to the correct tab on
download.  This is different from above.  If you pull_db but don't
switch to the proper tab, all results will be displayed.

Displays "It looks like nothing is here" if there are no documents for
the category.  Looks at Music for an example.